Cleanup: Simplify dependency around SSLNetVC and SNIConfig
