+++ iocore/net/SSLNextProtocolAccept.cc
@@ -22,6 +22,7 @@
  */
 
 #include "P_SSLNextProtocolAccept.h"
+#include "P_SSLNetVConnection.h"
 
 static void
 send_plugin_event(Continuation *plugin, int event, void *edata)
