+++ proxy/http/PreWarmManager.cc
@@ -25,7 +25,7 @@
 #include "PreWarmConfig.h"
 
 #include "HttpConfig.h"
-#include "P_SSLSNI.h"
+#include "SSLSNIConfig.h"
 
 #include "tscore/ink_time.h"
 #include "tscpp/util/PostScript.h"
